Filter out redundant references when preloading a `through` relation
When preloading a `through` relation, we first determine the
`middle_records`. These are the inhabitants of the relation we are reaching
through, and when there are multiple owners corresponding to a single record,
e.g. in a `belongs_to` case, they can be returned as multiple references to the
same Ruby object.

In such a situation, we do not want to calculate the members of the other side
of the relation for each such reference, because at best it would be redundant.
In the case of a `has_many`, we would bind multiple copies of the same row and
incorrectly inflate the number of entries in the preloaded association.

Fixes #37891