Avoid loading associations during `#reload` that were not originally loaded
When the logic for automatically reloading preloaded records was introduced, it
assumed that the presence of an entry in the association cache indicated that
that association had actually been loaded.  When a record is _written_ to an
association, e.g. by `parent.children.create` or `parent.children.scope.create`,
the association cache entry is created, but the original association records
have not yet been loaded.

This change alters that reload logic to only take effect when the association
had actually been loaded.